feat: exercise 4.21 solution

+++ part4/bloglist/controllers/blogs.js
@@ -55,6 +55,23 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  if(!request.token){
+    return response.status(401).json({ error: 'token misisng' })
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if(!blog){
+    return response.status(404).json({error: 'blog not found' })
+  }
+
+  if(blog.user.toString() !== decodedToken.id.toString()){
+    return response.status(403).json({ error: 'only the creator can delete this blog'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
